major update added new functionality. fixed some minor bugs.

+++ src/components/textarea.component.jsx
@@ -9,12 +9,16 @@ export default class TextArea extends React.Component {
                 {
                     value: 'Make a new task',
                     completed: false,
-                    id: 0
+                    id: 0,
+                    deleted: false,
+                    date: new Date()
                 },
                 {
                     value: 'Enter the website',
                     completed: true,
-                    id: 1
+                    id: 1,
+                    deleted: false,
+                    date: new Date()
                 }
             ],
             isValueExists: false,
@@ -32,18 +36,20 @@ export default class TextArea extends React.Component {
     
     }
 
-    handleOnClick() {
+    handleOnClick(e) {
+        e.preventDefault()
         // if value is empty or it already exists console.log it
         let isValueExists = false;
         let isValueEmpty = false;
         this.state.tasks.map(task => {
-            if (task.value === this.state.value) {
+            if (task.value === this.state.value && task.completed === false) {
                 // it already exists
                 isValueExists = true
             }
             if (this.state.value === '') {
                 isValueEmpty = true
             }
+            return null
         })
         if (isValueExists || isValueEmpty) {
             if (isValueExists) {
@@ -55,7 +61,9 @@ export default class TextArea extends React.Component {
             const newTask = {
                 value: this.state.value,
                 completed: false,
-                id: this.state.tasks.length
+                deleted: false,
+                id: this.state.tasks.length,
+                date: new Date()
             }
             let newTasks = this.state.tasks;
             newTasks.push(newTask)
@@ -105,7 +113,7 @@ export default class TextArea extends React.Component {
         while (iteration) {
             if (newTasks[i].id === id) {
                 console.log(`task ${i} has been deleted`);
-                newTasks.splice(i, 1);
+                newTasks[i].deleted = true;
                 this.setState({
                     tasks: newTasks
                 })
@@ -117,9 +125,9 @@ export default class TextArea extends React.Component {
     }
 
     handleEnterPress(e) {
-        if (e.keyCode == 13 && e.shiftKey == false) {
-            e.preventDefault();
-            this.handleOnClick();
+        if (e.keyCode === 13 && e.shiftKey === false) {
+            e.preventDefault()
+            this.handleOnClick(e);
         }
     }
 
@@ -127,7 +135,7 @@ export default class TextArea extends React.Component {
         return (
             <div>
                 <div className="textArea ">
-                    <form onSubmit={this.handleOnClick}>
+                    <form onSubmit={this.handleOnClick} className='general-form'>
                     <textarea onKeyDown={this.handleEnterPress} onChange={(e) => this.handleChange(e)} placeholder= 'Enter any activity' name="" id="" cols="30" rows="2"></textarea>
                     <button type='submit'>Submit</button>
                     </form>
@@ -138,7 +146,7 @@ export default class TextArea extends React.Component {
                     </div>
                 <div className="columns">
                     <div className='list'>
-                        Your tasks:
+                        <p className='list-label'>Your tasks:</p>
                         <ul className='list'>
                             {(
                             this.reverseArray(this.state.tasks).map(task => {
@@ -149,22 +157,27 @@ export default class TextArea extends React.Component {
                                         <button onClick={() => this.completeTask(task.id)}className='do-action-button'><i className="fas fa-check-square"></i></button>
                                     </div>
                                     </li>
+                                } else {
+                                    return null;
                                 }
                             }))}
                         </ul>
                     </div>
                     <div className='list'>
-                        Completed tasks:
+                        <p className='list-label'>Completed tasks:</p>
                         <ul className='list'>
                             { (
                                 this.state.tasks.map(task => {
-                                    if (task.completed) {
+                                    if (task.completed && task.deleted === false) {
                                         return <li key={task.id}>
                                         <div className='action'>
                                             <p className='goal'>{task.value}</p>
                                             <button onClick={() => this.deleteTask(task.id)} className='do-action-button'><i className="fas fa-trash-alt"></i></button>
                                         </div>
                                         </li>
+                                    } 
+                                     else {
+                                        return null;
                                     }
                                 })
                             )}
